Update int values

+++ AddressableLEDStrip.cpp
@@ -2,7 +2,7 @@
 #include "AddressableLEDStrip.h"
 
 // Create the LED strip object with the clock pin, serial data pin and the number of LEDs in the strip.
-AddressableLEDStrip::AddressableLEDStrip(uint8_t clk_pin, uint8_t sdi_pin, uint16_t num_leds)
+AddressableLEDStrip::AddressableLEDStrip(int clk_pin, int sdi_pin, int num_leds)
 {
 
   allow_out_of_bounds = false;
@@ -56,7 +56,7 @@ void AddressableLEDStrip::set_led(uint16_t index, int red, int green, int blue)
 
 // Set all the LEDs to the same RGB value.
 // This does not write them to the strip.
-void AddressableLEDStrip::set_all_leds(int8_t red, int8_t green, int8_t blue)
+void AddressableLEDStrip::set_all_leds(int red, int green, int blue)
 {
   for(int i = 0 ; i < strip_len ; i++) {
     set_led(i, red, green , blue);
@@ -111,7 +111,7 @@ void AddressableLEDStrip::_send_color_bits(int value)
 }
 
 // Get the number of LED that are in this strip (or at least what you set the length to be).
-uint16_t AddressableLEDStrip::length()
+int AddressableLEDStrip::length()
 {
   return strip_len;
 }
